requirements.txt
numpy==1.18.5
matplotlib==3.2.2
scipy==1.5.0
pandas==1.0.5
torch==1.7.0
pycocotools==2.0.2
scikit_image==0.17.2
torchvision==0.8.1
Pillow==8.0.1
SimpleITK==2.0.2
skimage==0.0
